# Optimize Status#permitted_for 500x (account timeline)